Split express example into text-only and request-construction variants
